Copy over changes from Joyce's commit - Use constants rather than hardcoded plugin names - Add missing types - Rename files that don't use JSX to be lowercase TS files - Add back React Router and Navigation link to About page (got nuked in squash) - Removed FeedPage (in lieu of Jakub's FeedView)
