+++ src/App.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { SharedStateProvider } from "./state";
+import { SharedStateProvider } from "./shared/state";
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import FeedView from "./components/FeedView";
 import Footer from "./components/Footer";
@@ -7,7 +7,7 @@ import Home from "./components/Home";
 import Navigation from "./components/Navigation";
 import Results from "./components/Results";
 import About from "./components/About";
-import { feedPageParameter } from "./shared/Constants";
+import { feedPageParameter } from "./shared/constants";
 
 class App extends React.Component {
 	render(): JSX.Element {
@@ -16,7 +16,11 @@ class App extends React.Component {
 				<Router>
 					<Navigation />
 					<Switch>
-						<Route path="/results/:id" component={FeedView} />
+						<Route path="/about" component={About} />
+						<Route
+							path={`/results/:${feedPageParameter}`}
+							component={FeedView}
+						/>
 						<Route path="/results" component={Results} />
 						<Route path="/" component={Home} />
 					</Switch>
