--- src/shared/Constants.ts
@@ -1,44 +1,0 @@
-export const dircopyPluginName = "pl-dircopy";
-export const med2ImgPluginName = "pl-med2img";
-export const infantFSPluginName = "pl-fshack-infant";
-
-export const feedPageParameter = "feedId";
-
-export const statsWithTableStructure = [
-	"aseg.stats",
-	"lh.aparc.a2009s.stats",
-	"lh.aparc.DKTatlas.stats",
-	"lh.aparc.pial.stats",
-	"lh.aparc.stats",
-	"lh.BA_exvivo.stats",
-	"lh.BA_exvivo.thresh.stats",
-	"lh.w-g.pct.stats",
-	"rh.aparc.a2009s.stats",
-	"rh.aparc.DKTatlas.stats",
-	"rh.aparc.pial.stats",
-	"rh.aparc.stats",
-	"rh.BA_exvivo.stats",
-	"rh.BA_exvivo.thresh.stats",
-	"rh.w-g.pct.stats",
-	"wmparc.stats",
-];
-
-export const keyToTextMap = {
-	id: "ID",
-	title: "Title",
-	previous_id: "Previous ID",
-	plugin_id: "Plugin ID",
-	plugin_name: "Plugin name",
-	plugin_version: "Plugin version",
-	pipeline_inst: "Pipeline instance",
-	feed_id: "Feed ID",
-	start_date: "Start date",
-	end_date: "End date",
-	status: "Status",
-	owner_username: "Owner username",
-	compute_resource_identifier: "Compute resource identifier",
-	cpu_limit: "CPU limit",
-	memory_limit: "Memory limit",
-	number_of_workers: "Number of workers",
-	gpu_limit: "GPU limit",
-};
