+++ src/shared/constants.ts
@@ -1,0 +1,5 @@
+export const dircopyPluginName = "pl-dircopy";
+export const med2ImgPluginName = "pl-med2img";
+export const infantFSPluginName = "pl-fshack-infant";
+
+export const feedPageParameter = "feedId";
