+++ src/shared/state.ts
